debug: update imports and move buttonStyle to App to troubleshoot potential circular dependency
