+++ src/App.jsx
@@ -8,6 +8,12 @@ import { useShoppingLists } from './api';
 import { ToastContainer } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 
+export const buttonStyle = {
+	color: 'white',
+	width: '15em',
+	fontSize: '1.5rem',
+};
+
 export function App() {
 	/**
 	 * This custom hook takes the path of a shopping list
